% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_id}
\alias{add_id}
\title{Add an Id to the snippet}
\usage{
add_id(snip)
}
\arguments{
\item{snip}{The snippet}
}
\value{
Id of the file (either the one already present, or a new one)
}
\description{
Add an Id to the snippet
}
