% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_section}
\alias{add_section}
\title{Add a section to a snip}
\usage{
add_section(snip, section, x)
}
\arguments{
\item{snip}{A snippet}

\item{section}{Section header}

\item{x}{Content}
}
\value{
An updated snippet
}
\description{
The function does not verify if the section exists already
}
