% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_headers}
\alias{extract_headers}
\title{Extract headers from a snippet}
\usage{
extract_headers(snip, code_only = T)
}
\arguments{
\item{snip}{The snippet}

\item{code_only}{If true, only the ones with code}
}
\value{
Character vector of headers
}
\description{
Extract headers from a snippet
}
