% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_info}
\alias{extract_info}
\title{Extract section info from snippet}
\usage{
extract_info(snip, type, filter_comments = TRUE)
}
\arguments{
\item{snip}{A snippet}

\item{type}{Section to extract (Name, Tags, Description, Packages, etc)}

\item{filter_comments}{Bool}
}
\value{
The information as character vector
}
\description{
Extract section info from snippet
}
