% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_titles}
\alias{extract_titles}
\title{Extract code section title from string}
\usage{
extract_titles(x)
}
\arguments{
\item{x}{Header(s)}
}
\value{
Title(s)
}
\description{
Gets rid of leading and trailing pound signs
}
