% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_filename}
\alias{get_filename}
\title{Filename for new snip}
\usage{
get_filename(n = 10L)
}
\arguments{
\item{n}{Recursive counter}
}
\value{
path_to_file
}
\description{
Create filename to save a new snip
}
