% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_id}
\alias{get_id}
\title{Get the snip id from location}
\usage{
get_id(fname)
}
\arguments{
\item{fname}{Character string with snip location or snip name}
}
\value{
The id as character
}
\description{
Get the snip id from location
}
