% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_modal_page}
\alias{get_modal_page}
\title{Create UI with inputs for current page in modal for creating snippets}
\usage{
get_modal_page(rv)
}
\arguments{
\item{rv}{reactiveValues()}
}
\value{
UI for the modal
}
\description{
First page is the meta info (Name, Tags, etc). Second to fifth pages are code
}
