% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_path_to_folder}
\alias{get_path_to_folder}
\title{Interactively get a path to folder}
\usage{
get_path_to_folder(caption)
}
\arguments{
\item{caption}{Window caption}
}
\value{
Path
}
\description{
Interactively get a path to folder
}
