% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_fname}
\alias{make_fname}
\title{Create full path to file (in saved snippets folder) given an id}
\usage{
make_fname(id)
}
\arguments{
\item{id}{Id of a snippet}
}
\value{
File path for snippet
}
\description{
Create full path to file (in saved snippets folder) given an id
}
