% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_snip}
\alias{read_snip}
\title{Read snippet specified by row number of snip_view}
\usage{
read_snip(i)
}
\arguments{
\item{i}{Row number}
}
\value{
Character vector with snippet
}
\description{
Read a snippet from file. Specify snippet with row number in snip_view
}
