% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_invalid_headers}
\alias{remove_invalid_headers}
\title{Remove invalid headers}
\usage{
remove_invalid_headers(snip)
}
\arguments{
\item{snip}{A snippe}
}
\value{
A snippet with invalid headers removed
}
\description{
Removes trailing pound signs from any header that is not one of the main ones, or an Item_ one
}
