% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_delete}
\alias{snip_delete}
\title{Delete a snippet}
\usage{
snip_delete(i = NULL, id = NULL)
}
\arguments{
\item{i}{The row number as indicated by snip_view()}

\item{id}{The snippet's Id}
}
\value{

}
\description{
Use either the snippet Id or its row number from snip_view to select the snippet to delete.
}
\examples{
snip_delete(i=1)
snip_delete(id=7059297)
}
