% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_edit}
\alias{snip_edit}
\title{Edit a new snippet}
\usage{
snip_edit(i = NULL, id = NULL)
}
\value{

}
\description{
Edit an existing snippet.
Edit the file. Once done, save the file and save the changes with snip_save().
}
