% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_export}
\alias{snip_export}
\title{Exports snippets a zip file}
\usage{
snip_export(path = NULL, fname = "snippets.zip")
}
\arguments{
\item{path}{Path to directory. Leave empty to choose one interactively.}

\item{fname}{Name of the exported zip}
}
\description{
Choose a directory interactively (or specify directly with path argument).
}
