% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_fix}
\alias{snip_fix}
\title{Attempt to fix snippie by reloading snippet information}
\usage{
snip_fix(play_it_safe = TRUE)
}
\arguments{
\item{play_it_safe}{Bool, if TRUE (default) creates a zip of all saved snippets}
}
\value{
data.frame of available snippets
}
\description{
Attempt to fix snippie by reloading snippet information
}
\examples{
snip_fix()
}
