% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_folder}
\alias{snip_folder}
\title{Select a folder full op undocumented snippets to add to the snippie library}
\usage{
snip_folder(path = NULL, recursive = TRUE, extensions = NULL)
}
\arguments{
\item{path}{Path to directory. Leave empty to choose one interactively.}

\item{recursive}{Boolean}

\item{extensions}{A character vector of extensions, eg c('py', 'R', 'txt')}
}
\value{

}
\description{
All files in a folder can be added to snippie in one go.
snippie will use the file name as Title and add an Id.
}
\examples{
snip_folder(recursive=TRUE, extensions=c('txt', 'R'))
}
