% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_new}
\alias{snip_new}
\title{Create a new snippet}
\usage{
snip_new()
}
\value{

}
\description{
Create a new snippet.
This will create a template to fill in.
Once done, save the file and add it to your snippets with snip_save().
}
