% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{snip_view}
\alias{snip_view}
\title{View available snips}
\usage{
snip_view(n, exact = F, ...)
}
\arguments{
\item{n}{Can be used to view and edit all information regarding a specific snippet.
n refers to the row.number returned by snip_view()}

\item{exact}{How precise the subsetting should be. Exact==TRUE is not fully exact (yet).}

\item{...}{Subset on any of the columns, eg, snip_view(Package=data.table, Tags=column)}
}
\value{
The snippets as data.frame
}
\description{
Without arguments, all snippets are shown. Select a subset using ...
}
\examples{
snip_view(Package=data.table, Tags=column)
snip_view(p=plotly)
}
