% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_d}
\alias{update_d}
\title{Update data.frame with available snippets}
\usage{
update_d(fname, snip = NULL, id = NULL)
}
\arguments{
\item{fname}{Filename where latest snip got saved}
}
\value{
invisible(TRUE) when successful
}
\description{
Update data.frame with available snippets
}
