% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_snip}
\alias{update_snip}
\title{Function wrapper to update snippet}
\usage{
update_snip(fname, f)
}
\arguments{
\item{fname}{File name of snippet}

\item{f}{Function to apply to the snippet}
}
\description{
Function wrapper to update snippet
}
