% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{update_snip_info}
\alias{update_snip_info}
\title{Update snippet info in rv}
\usage{
update_snip_info(rv, input)
}
\arguments{
\item{rv}{reactiveValues}

\item{input}{input}
}
\value{

}
\description{
Whenever the user changes the page inside the modal for creating snippets the info needs to be updated.
Since the code input pages will use the same input bindings.
}
