% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate_i_and_id}
\alias{validate_i_and_id}
\title{Validate i and id arguments}
\usage{
validate_i_and_id(i, id)
}
\arguments{
\item{i}{The row index for the snippet}

\item{id}{The id of the snippet}
}
\value{
Either calls stop or returns list with i and id
}
\description{
Functions that use either i (row index) or id (snip id) can use
this function to validate correct input
}
