% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate_snip}
\alias{validate_snip}
\title{Validate snippet before saving}
\usage{
validate_snip(fname, snip = NULL)
}
\arguments{
\item{fname}{Location of snippet}
}
\value{
TRUE if successful, otherwise it calls stop
}
\description{
Validate snippet before saving
}
